% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{adj_to_inc_reduced}
\alias{adj_to_inc_reduced}
\title{incidence matrix from adjacency matrix in reduced form}
\usage{
adj_to_inc_reduced(adj)
}
\arguments{
\item{adj}{an adjacency matrix}
}
\value{
an incidence matrix
}
\description{
takes an adjacency matrix and generates the incidencematrix
}
