% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphUtils.R
\name{combine_roll_call}
\alias{combine_roll_call}
\title{combine_roll_call}
\usage{
combine_roll_call(roll_call1, roll_call2)
}
\arguments{
\item{roll_call}{a list of class roll call, assumed to bea generated using input_roll_call}

\item{p}{probability/frequency of mutation}
}
\value{
combine_roll_call
}
\description{
combine_roll_call
}
