% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphUtils.R
\name{compute_accuracy}
\alias{compute_accuracy}
\title{accuracy roll call predict}
\usage{
compute_accuracy(roll_call)
}
\arguments{
\item{roll_call}{a list of class roll call, assumed to bea generated using input_roll_call}
}
\value{
data frame of accuracy
}
\description{
accuracy call graph
}
