% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphUtils.R
\name{compute_homology_rank}
\alias{compute_homology_rank}
\title{homology rank (first betti number)}
\usage{
compute_homology_rank(roll_call)
}
\arguments{
\item{roll_call}{a list of class roll call, assumed to be generated using input_roll_call}
}
\value{
the first betti number of the graph
}
\description{
computes the main fitness function, the rank of the first homology group
}
