% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphUtils.R
\name{crossover_roll_call}
\alias{crossover_roll_call}
\title{crossover_roll_call}
\usage{
crossover_roll_call(roll_call1, roll_call2, type = c("join", "intersect",
  "pool"), pool_rate = 0.25)
}
\arguments{
\item{roll_call}{a list of class roll call, assumed to bea generated using input_roll_call}

\item{p}{probability/frequency of mutation}
}
\value{
crossover_roll_call
}
\description{
crossover_roll_call
}
