% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphUtils.R
\name{gen_adj_matrix}
\alias{gen_adj_matrix}
\title{initial adj matrix}
\usage{
gen_adj_matrix(d)
}
\description{
receives roll call data and creates an adj matrix
this matrix creates an adjacency matrix from roll call data that has perfect prediction
but usually has the highest possible dimension.
}
