% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphUtils.R
\name{gen_nodes}
\alias{gen_nodes}
\title{all nodes}
\usage{
gen_nodes(d)
}
\description{
creates the nodes of voters and issues for the roll_call class graph
}
\details{
the roll_call class is made of a graph which consists of a noder per voter and two
nodes for each issue voted representing the yes and no alternatives.
}
\examples{
data(onedim_voting)
gen_nodes(onedim_voting)
}
