% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphUtils.R
\name{input_roll_call}
\alias{input_roll_call}
\title{voting class constructor}
\usage{
input_roll_call(d)
}
\arguments{
\item{x}{a data.frame with each row representing a voter and each column the result of one roll call vote}
}
\value{
\itemize{
    \item \code{nodes} names of all the nodes
    \item \code{voter} the indexes of voters in \code{nodes}
    \item \code{Y} the indexes of 'yes' nodes for each issue in \code{nodes}
    \item \code{N} the indexes of 'no' nodes for each in \code{nodes}
}
}
\description{
defines the class \code{roll_call}
}
\details{
returns a list of class {roll_call} with elements
}
