% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphUtils.R
\name{min_hom}
\alias{min_hom}
\title{optimise graph with min homology}
\usage{
min_hom(roll_call, population_size = 250, n_generations = 50,
  survival_rate = 0.75, mutation_rate = 0.05, max_time = 300)
}
\arguments{
\item{roll_call}{a list of class roll call, assumed to bea generated using input_roll_call}

\item{population_size}{xx}

\item{n_generations}{xx}

\item{survival_rate}{xx}

\item{max_time}{xx}

\item{mutation_tate}{xx}
}
\value{
optimized graph
}
\description{
optimise
}
