% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphUtils.R
\name{mutate_roll_call}
\alias{mutate_roll_call}
\title{mutate}
\usage{
mutate_roll_call(roll_call, p = 0.05)
}
\arguments{
\item{roll_call}{a list of class roll call, assumed to bea generated using input_roll_call}

\item{p}{probability/frequency of mutation}
}
\value{
mutated roll call
}
\description{
mutate
}
