% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{onedim_voting}
\alias{onedim_voting}
\title{onedim_voting}
\format{This dataset gives an example of a truly one-dimensional voting system
onedim_voting <- data.frame(
    row.names = c("A", "B", "C", "D", "E"),
    I1 = c(1, 1, 0, 0, 0),
    I2 = c(1, 1, 1, 0, 0),
    I3 = c(1, 1, 1, 1, 0),
    I4 = c(0, 1, 1, 1, 1),
    I5 = c(1, 1, 1, 1, 1)
)}
\usage{
onedim_voting
}
\keyword{datasets}
