% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphUtils.R
\name{plot.roll_call}
\alias{plot.roll_call}
\title{plot roll call graph}
\usage{
\method{plot}{roll_call}(roll_call, method = c("networkD3", "igraph"), ...)
}
\arguments{
\item{roll_call}{a list of class roll call, assumed to be generated using input_roll_call}
}
\description{
plot roll call graph
}
