% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphUtils.R
\name{predict.roll_call}
\alias{predict.roll_call}
\title{predict roll call predict}
\usage{
\method{predict}{roll_call}(roll_call)
}
\arguments{
\item{roll_call}{a list of class roll call, assumed to be generated using input_roll_call}
}
\value{
data frame of predictions
}
\description{
predict call graph
}
