% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{twodim_voting}
\alias{twodim_voting}
\title{twodim_voting}
\format{This dataset gives an example of a truly two-dimensional voting system \code{
twodim_voting <- data.frame(
    row.names = c("A", "B", "C", "D", "E", "F", "G"),
    I1 = c(1, 1, 0, 0, 0, 0, 0),
    I2 = c(1, 1, 1, 0, 0, 0, 0),
    I3 = c(0, 0, 1, 1, 0, 0, 1),
    I4 = c(0, 1, 1, 1, 0, 0, 1),
    I5 = c(1, 1, 1, 1, 0, 0, 0)
) }}
\usage{
twodim_voting
}
\keyword{datasets}
