% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphUtils.R
\name{gen_adj_matrix}
\alias{gen_adj_matrix}
\title{initial incidence matrix}
\usage{
gen_adj_matrix(d)
}
\description{
receives roll call data and creates an incidence matrix
}
\details{
the incidence matrix \code{E} has entry \code{e_{ij}} equal to
\itemize{
    \code{1} if an edge starts in node \code{i} and finishes in \code{j}
    \code{-1} if an edge ends in node \code{i} and finishes in \code{j}
    \code{0} if nodes \code{i} and \code{j} are not connected
}
this matrix creates an incidence matrix from roll call data that has perfect prediction
but usually has the highest possible dimension.
}
\examples{
data(onedim_voting)
create_incidence_matrix(onedim_voting)
}
